+++ sharding-core/sharding-core-parse/sharding-core-parse-postgresql/src/main/java/org/apache/shardingsphere/core/parse/PostgreSQLParserEngine.java
@@ -19,23 +19,23 @@ package org.apache.shardingsphere.core.parse;
 
 import org.antlr.v4.runtime.CharStreams;
 import org.antlr.v4.runtime.CommonTokenStream;
-import org.apache.shardingsphere.core.parse.autogen.MySQLStatementLexer;
-import org.apache.shardingsphere.core.parse.spi.ShardingParseEngine;
+import org.apache.shardingsphere.core.parse.autogen.PostgreSQLStatementLexer;
+import org.apache.shardingsphere.core.parse.spi.SQLParserEngine;
 
 /**
- * Sharding parse engine for H2.
+ * SQL parser engine for PostgreSQL.
  *
  * @author zhangliang
  */
-public final class H2ShardingParseEngine implements ShardingParseEngine {
+public final class PostgreSQLParserEngine implements SQLParserEngine {
     
     @Override
     public String getDatabaseType() {
-        return "H2";
+        return "PostgreSQL";
     }
     
     @Override
-    public MySQLParser createSQLParser(final String sql) {
-        return new MySQLParser(new CommonTokenStream(new MySQLStatementLexer(CharStreams.fromString(sql))));
+    public PostgreSQLParser createSQLParser(final String sql) {
+        return new PostgreSQLParser(new CommonTokenStream(new PostgreSQLStatementLexer(CharStreams.fromString(sql))));
     }
 }
