+++ sharding-core/sharding-core-parse/sharding-core-parse-postgresql/src/main/resources/META-INF/services/org.apache.shardingsphere.core.parse.spi.SQLParserEngine
@@ -15,4 +15,4 @@
 # limitations under the License.
 #
 
-org.apache.shardingsphere.core.parse.OracleShardingParseEngine
+org.apache.shardingsphere.core.parse.PostgreSQLParserEngine
