+++ sharding-core/sharding-core-parse/sharding-core-parse-sqlserver/src/main/resources/META-INF/services/org.apache.shardingsphere.core.parse.spi.SQLParserEngine
@@ -15,5 +15,4 @@
 # limitations under the License.
 #
 
-org.apache.shardingsphere.core.parse.MySQLShardingParseEngine
-org.apache.shardingsphere.core.parse.H2ShardingParseEngine
+org.apache.shardingsphere.core.parse.SQLServerParserEngine
